Remove unsupportable intrinsic sizes from intrinsic_def.json
The intrinsics serve a dual duty:
  1. They are used to implement regular instructions.
  2. They are used to implement vector instrinsics.

And there's a dilemma: 8bit and 16bit sizes are not supported in #1 but
are needed for #2.

Add tiny hack to gen_intrinsics.py as short-term solution. If we would
need more such intrinsics we may need to develop and add some special
markup to our gen_intrinsics.py files.

Bug: 401585338

Test: m berberis_all

Change-Id: I03862c1d9d93a14ee3788dcf2fc8622831102fef