Rename Div and Rem intrinsics.
Most intrinsics either don't have the name of x86 instruction (like
sh1adduw) or do approximately the same thing (like add).

But Div is exception: on Risc-V and on x86 Div behave differently and
we have macro-instruction that handles that difference.

Rename Div intrinsic to RiscVDiv (and Rem to RiscVRem for consistency).

Bug: 399130034

Test: m befrberis_all

Change-Id: Ie0d360198d566cec405f78e8271b5544aea80ac1