Add GenSSA<…> code generator.
It encapsulates allocation of temprorary registers and is supposed to
produce SSA insructions in the future (with step of allocation of
temporary registers for use_def arguments to be moved closer to the
machine code generation time).

Convert Gen<…> in frontend.h for testing (tricky cases are redirected to
builder_.Gen<…> instead).

Test: m berberis_all

Change-Id: Id9476d1b1658cd5e479426b6656a37b227c9083c