Rename Apply and Generate in tuple_processing.h
These names look very confusing when compared with std::apply, std::generate
and std::generator.

There are minor confusion with std::map, too, but Map operation is very
well-known, thanks to the use in JavaScript, Python and many other
languages.

ForEach looks similar to std::for_each and analogous functions in many
other languages and Produce shouldn't cause any associations with parts
of standard library—we need it to produce constexpr reasults in a “piecemeal”
fashion, it's not really needed when working in purely imperative code.

Test: m berberis_all

Bug: 457703675

Change-Id: I2af5acd640f8580d46e5002fa8c8b4b6a53e80d3