backend: fix invalid vector use in dead code removal
`bit_set_.clear()` sets the size of the vector here to 0, making it
invalid to use. When built with libcxx hardening, host tests crash on
subsequent `Set()`s.

Rename `Clear()` to be more specific about what the desirable behavior
is, and make it zero out the bitset instead of truncating it to 0 elems.

Bug: 465498284
Test: atest --host-unit-test-only berberis_host_tests
Change-Id: I744d2df4afefe520a916eaec7278aadb4c81675c