Make MetaType functions with arguments more versatile.
With one argument we don't have a choice: we only need a type for these
and thus `static constexpr` function is logical.

With more than one argument situation is less obvious: when we have
precisely two types second argument could be type or meta-type and if we
have arbitrary number of arguments we may want to accept tuple of meta-types,
because that's what kTupleMetaTypes is producing.

Test: m berberis_all

Change-Id: Iaeaed314deb8862d6ba4ffbf4daa5715a2fb769f