Remove friends from WrappedFloatType definition.
FLoat16/Float32/Float64 do a double-duty of providing a container for
moving float values (whcih we need in various places) and also providing
safe operations (which is impossible to do in C++, to guarantee corrent
handling of NaNs and rounding we need inline assembler) — but these are
only needed in intrinsics.

Thankfully we guarantee that types are layout-compatible (and can even
test that when https://github.com/llvm/llvm-project/issues/48204 would
be fixed and libc++ would be C++20 compatible).

Test: m berberis_all

Change-Id: Ic71fd6d05082a08c1eb52d7ba2a56d0f1e070dce