Remove Float-related hack
We needed to define functions that way to make them compatible with
definitions of Float16/32/64 in intrinsics when NarroType and WideType
was defined in the bate.

Now we no longer need it and can just rely on the fact that
type_traits.h includes both definition of int types and float types.

Test: m berberis_all

Change-Id: Ibe6ab1528f3d4041abefd9fc6fc0812699c05b4b